로또 3단계 피드백 반영중 3차
