+++ src/main/java/raffle/lotto/play/PlayLotto.java
@@ -4,6 +4,7 @@ import raffle.lotto.LottoMachine;
 import raffle.lotto.money.Money;
 import raffle.lotto.validator.LottoNumberValidator;
 import raffle.lotto.win.LottoResult;
+import raffle.lotto.win.LottosCreation;
 import raffle.view.InputView;
 import raffle.view.ResultView;
 
@@ -11,7 +12,8 @@ public class PlayLotto {
     public static void main(String[] args) {
         ResultView resultView = new ResultView();
         Money money = new Money(InputView.buyLotto() , InputView.lottoAmount());
-        LottoMachine lottoMachine = new LottoMachine(money ,InputView.getLottoNumber(money.getManualCount()), new LottoNumberValidator());
+        LottoMachine lottoMachine =
+                new LottoMachine(money ,InputView.getLottoNumber(money.getManualCount()), new LottoNumberValidator(), new LottosCreation());
         resultView.buyLotto(lottoMachine.getLottos(), money.getManualCount());
 
         LottoResult lottoResult = lottoMachine.winLotto(InputView.lastWeekLotto(), InputView.bonusLotto());
