+++ src/main/java/raffle/lotto/win/LottosGenerator.java
@@ -5,7 +5,6 @@ import raffle.lotto.money.Money;
 
 import java.util.List;
 
-@FunctionalInterface
 public interface LottosGenerator {
     List<Lotto> generate(Money money);
 }
