+++ src/test/java/raffle/lotto/LottoMachineTest.java
@@ -1,10 +1,9 @@
 package raffle.lotto;
 
-import org.junit.Before;
 import org.junit.Test;
 import raffle.lotto.money.Money;
 import raffle.lotto.validator.LottoNumberValidator;
-import raffle.view.InputView;
+import raffle.lotto.win.LottosCreation;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -15,30 +14,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 public class LottoMachineTest {
 
-    LottoMachine lottoMachine;
-
-    @Before
-    public void setUp(){
-        List<Lotto> lottos = new ArrayList<>();
-        lottos.add(new Lotto(Arrays.asList(1, 6, 5, 44, 22, 15)));
-
-        Money money = new Money(3 , 13000);
-
-        lottoMachine = new LottoMachine(money, lottos, new LottoNumberValidator());
-    }
-
     @Test
     public void 로또입력확인() {
-        assertThat(lottoMachine.getLottos()).hasSize(13);
-    }
+        List<Lotto> lottos = new ArrayList<>();
+        lottos.add(new Lotto(Arrays.asList(1, 6, 5, 44, 22, 45)));
+        lottos.add(new Lotto(Arrays.asList(1, 2, 5, 44, 22, 15)));
+        lottos.add(new Lotto(Arrays.asList(1, 6, 30, 44, 32, 15)));
 
-    @Test(expected = RuntimeException.class)
-    public void 로또_수동_잘못된_숫자() {
-        List<Lotto> notLottos = new ArrayList<>();
-        notLottos.add(new Lotto(Arrays.asList(1, 6, 5, 44, 22, 77)));
         Money money = new Money(3 , 13000);
 
-        lottoMachine = new LottoMachine(money, notLottos, new LottoNumberValidator());
+        LottoMachine lottoMachine = new LottoMachine(money, lottos, new LottoNumberValidator(), new LottosCreation());
+        assertThat(lottoMachine.getLottos()).hasSize(13);
     }
 
 }
